+++ src/alley/wp/match-blocks.php
@@ -18,6 +18,7 @@ use Alley\WP\Validator\Block_Name;
 use Alley\WP\Validator\Block_Offset;
 use Alley\WP\Validator\Block_InnerBlocks_Count;
 use Alley\WP\Validator\Nonempty_Block;
+use Laminas\Validator\ValidatorInterface;
 
 /**
  * Match blocks within the given content.
@@ -48,6 +49,7 @@ use Alley\WP\Validator\Nonempty_Block;
  *    }
  *    @type bool                      $count             Return the number of found blocks instead of the set.
  *    @type bool                      $has_innerblocks   Return only blocks that have, or don't have, inner blocks.
+ *    @type ValidatorInterface        $is_valid          Match blocks that pass the given validator.
  *    @type bool                      $flatten           Recursively descend into inner blocks, test each one against the
  *                                                       criteria, and count each towards totals. Default false.
  *    @type int                       $limit             Extract at most this many blocks. Default `-1`, or no limit.
@@ -77,6 +79,7 @@ function match_blocks( $source, $args = [] ) {
 			'count'             => false,
 			'flatten'           => false,
 			'has_innerblocks'   => null,
+			'is_valid'          => null,
 			'limit'             => -1,
 			'name'              => '',
 			'nth_of_type'       => null,
@@ -192,6 +195,10 @@ function match_blocks( $source, $args = [] ) {
 					],
 				),
 			);
+		}
+
+		if ( $args['is_valid'] instanceof ValidatorInterface ) {
+			$validator->attach( $args['is_valid'] );
 		}
 	} catch ( \Exception $exception ) {
 		return $error;