+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 
 ### Added
 
-- Passing a single block will return matches within its inner blocks.
+- Passing a single block instance will return matches within its inner blocks.
 
 ## 1.0.1