+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 ### Added
 
 - Passing a single block instance will return matches within its inner blocks.
+- `with_innerhtml` parameter for matching blocks by their inner HTML.
 
 ## 1.0.1