+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 ### Added
 
 - `with_innerhtml` parameter for matching blocks by their inner HTML.
+- `CONTAINS` and `NOT CONTAINS` (case-sensitive), and `LIKE` and `NOT LIKE` (case-insensitive) operators to `attrs` parameter.
 - Passing a single block instance will return matches within its inner blocks.
 
 ## 1.0.1